refactor: please ignore this commit
Look, I'm trying really hard to maintain small, descriptive commits on this project, but this one ain't it chief. This is a mess. This was going to be a clean commit to add some unit tests but I ran into an issue with an earlier commit where I used lazy_static to create a static ref for the index handle, and after the struggle bus ride I just had of spurious test failures and other nonsense, I cannot be fucked to revert a commit.

State of the code: bad.
State of my mental health: also bad.
But still, we charge forward.